refactor: RequestHeader 내의 parse 함수를 제거 및 RequestHeaderParser 객체 생성
RequestHeader 클래스의 헤더를 파싱하는 메소드를 제거하고, RequestHeader의 내용을 파싱하는 객체 RequestHeaderParser를 생성하여 역할을 분담하도록 하였다.
