+++ src/main/java/header/RequestHeader.java
@@ -9,8 +9,6 @@ import java.lang.reflect.Field;
 import java.util.StringTokenizer;
 
 public class RequestHeader {
-    private static final Logger logger = LoggerFactory.getLogger(RequestHeader.class);
-
     private String method;
     private String path;
     private String protocol;
@@ -38,42 +36,6 @@ public class RequestHeader {
 
     public static RequestHeader of(String method, String path, String Host){
         return new RequestHeader(method, path, Host);
-    }
-
-    public static RequestHeader parseHeader(BufferedReader br) throws IOException {
-        String requestLine = br.readLine();
-        StringTokenizer st = new StringTokenizer(requestLine, " ");
-
-        String method = st.nextToken();
-        String path = st.nextToken();
-        String protocol = st.nextToken();
-
-        RequestHeader requestHeader = of(method, path, protocol);
-
-        while(!(requestLine = br.readLine()).isEmpty()){
-            String[] requestLineSplit = requestLine.split(": ");
-
-            String headerName = requestLineSplit[0].replace("-", "");
-            String headerValue = requestLineSplit[1];
-
-            setHeaderValue(requestHeader, headerName, headerValue);
-        }
-
-        return requestHeader;
-    }
-
-    private static void setHeaderValue(RequestHeader requestHeader, String fieldName, String fieldValue){
-        Class<RequestHeader> requestHeaderClass = RequestHeader.class;
-
-        try{
-            Field field = requestHeaderClass.getDeclaredField(fieldName);
-            field.setAccessible(true);
-
-            field.set(requestHeader, fieldValue);
-        } catch (NoSuchFieldException | IllegalAccessException e) {
-            logger.error(e.getMessage());
-            throw new RuntimeException(e);
-        }
     }
 
     @Override
